// new-app/src/entities/parameter/model/types.ts
export type ParameterSummaryData = {
  isActive: boolean;
  isFinished: boolean;
  productCode: string;
  productName: string;
  batchName: string;
  date: Date;
  shift: number;
  plan: number;
};